requirements.txt
# Basic
jupyter==1.0.0
python-dotenv==1.0.0
numpy==1.25.0
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
openai==0.27.8
tiktoken==0.4.0

# Function Call~
requests==2.31.0
beautifulsoup4==4.12.2
google-api-python-client==2.93.0

# Llama Index
llama_index==0.7.5
flask==2.3.2

# LangChain
langchain==0.0.230
pypdf==3.12.1
hnswlib==0.7.0
pydantic==1.10.8
chromadb==0.3.26
unstructured==0.8.1
wikipedia==1.4.0

# Streamlit
streamlit==1.24.1
